Move files to separate abstract classes and implementation in adapters
