+++ src/fastlife/adapters/jinjax/__init__.py
@@ -1,0 +1,7 @@
+"""Template renderer."""
+
+from .renderer import JinjaxTemplateRenderer
+
+__all__ = [
+    "JinjaxTemplateRenderer",
+]
