+++ src/fastlife/adapters/jinjax/widgets/boolean.py
