+++ src/fastlife/adapters/jinjax/widgets/checklist.py
