+++ src/fastlife/components/A.jinja
@@ -8,7 +8,7 @@
   id: Annotated[str | None, "unique identifier of the element."] = None,
   class_: Annotated[
     str | None,
-    "css class for the node, defaults to :attr:`fastlife.templates.renderer.constants.Constants.A_CLASS`."
+    "css class for the node, defaults to :attr:`fastlife.templates.constants.Constants.A_CLASS`."
   ] = None,
   hx_target: Annotated[
     str,
