+++ src/fastlife/components/Button.jinja
@@ -10,7 +10,7 @@
   class_: Annotated[
     str | None,
     "css class for the node, defaults to "
-    ":attr:`fastlife.templates.renderer.constants.Constants.BUTTON_CLASS`."
+    ":attr:`fastlife.templates.constants.Constants.BUTTON_CLASS`."
   ] = None,
   name: str = "action",
   value: str = "submit",
