+++ src/fastlife/components/Checkbox.jinja
@@ -7,7 +7,7 @@
   class_: Annotated[
     str | None,
     "css class for the node, defaults to "
-    ":attr:`fastlife.templates.renderer.constants.Constants.CHECKBOX_CLASS`."
+    ":attr:`fastlife.templates.constants.Constants.CHECKBOX_CLASS`."
   ] = None,
   value: Annotated[str | None, "http submitted value if the checkbox is checked"] = None,
   checked: Annotated[bool, "Initialized the checkbox as ticked"] = False,
