+++ src/fastlife/components/Details.jinja
@@ -19,7 +19,7 @@
   class_: Annotated[
     str | None,
     "css class for the node, defaults to "
-    ":attr:`fastlife.templates.renderer.constants.Constants.DETAILS_CLASS`."
+    ":attr:`fastlife.templates.constants.Constants.DETAILS_CLASS`."
   ] = None,
   open: Annotated[bool, "open/close state."] = True,
 #}
