+++ src/fastlife/components/H1.jinja
@@ -4,7 +4,7 @@
   class_: Annotated[
     str | None,
     "css class for the node, defaults to "
-    ":attr:`fastlife.templates.renderer.constants.Constants.H1_CLASS`."
+    ":attr:`fastlife.templates.constants.Constants.H1_CLASS`."
   ] = None,
 #}
 <h1
