+++ src/fastlife/components/H2.jinja
@@ -4,7 +4,7 @@
   class_: Annotated[
     str | None,
     "css class for the node, defaults to "
-    ":attr:`fastlife.templates.renderer.constants.Constants.H2_CLASS`."
+    ":attr:`fastlife.templates.constants.Constants.H2_CLASS`."
   ] = None,
 #}
 <h2
