+++ src/fastlife/components/H6.jinja
@@ -4,7 +4,7 @@
   class_: Annotated[
     str | None,
     "css class for the node, defaults to "
-    ":attr:`fastlife.templates.renderer.constants.Constants.H6_CLASS`."
+    ":attr:`fastlife.templates.constants.Constants.H6_CLASS`."
   ] = None,
 #}
 <h6
