+++ src/fastlife/components/Input.jinja
@@ -9,7 +9,7 @@
   class_: Annotated[
     str | None,
     "css class for the node, defaults to "
-    ":attr:`fastlife.templates.renderer.constants.Constants.INPUT_CLASS`."
+    ":attr:`fastlife.templates.constants.Constants.INPUT_CLASS`."
   ] = None,
   aria_label: Annotated[str | None, "aria-label"] = None,
   placeholder: Annotated[
