+++ src/fastlife/components/Label.jinja
@@ -7,7 +7,7 @@
   class_: Annotated[
     str | None,
     "css class for the node, defaults to "
-    ":attr:`fastlife.templates.renderer.constants.Constants.LABEL_CLASS`."
+    ":attr:`fastlife.templates.constants.Constants.LABEL_CLASS`."
   ] = None,
 #}
 <label
