+++ src/fastlife/components/P.jinja
@@ -6,7 +6,7 @@
   class_: Annotated[
     str | None,
     "css class for the node, defaults to "
-    ":attr:`fastlife.templates.renderer.constants.Constants.P_CLASS`."
+    ":attr:`fastlife.templates.constants.Constants.P_CLASS`."
   ] = None,
 #}
 <p
