+++ src/fastlife/components/Select.jinja
@@ -7,7 +7,7 @@ id: Annotated[str | None, "unique identifier of the element."] = None,
 class_: Annotated[
   str | None,
   "css class for the node, defaults to "
-  ":attr:`fastlife.templates.renderer.constants.Constants.SELECT_CLASS`."
+  ":attr:`fastlife.templates.constants.Constants.SELECT_CLASS`."
 ] = None,
 multiple: Annotated[bool, "Mark as multiple"] = False,
 #}
