+++ src/fastlife/components/Summary.jinja
@@ -6,7 +6,7 @@
   class_: Annotated[
     str | None,
     "css class for the node, defaults to "
-    ":attr:`fastlife.templates.renderer.constants.Constants.SUMMARY_CLASS`."
+    ":attr:`fastlife.templates.constants.Constants.SUMMARY_CLASS`."
   ] = None,
   open: Annotated[bool, "Open or collapse the content of the details."] = True
 #}
