+++ src/fastlife/components/Table.jinja
@@ -4,7 +4,7 @@
   class_: Annotated[
     str | None,
     "css class for the node, defaults to "
-    ":attr:`fastlife.templates.renderer.constants.Constants.TABLE_CLASS`."
+    ":attr:`fastlife.templates.constants.Constants.TABLE_CLASS`."
   ] = None,
 #}
 <table
