+++ src/fastlife/components/Td.jinja
@@ -4,7 +4,7 @@
   class_: Annotated[
     str | None,
     "css class for the node, defaults to "
-    ":attr:`fastlife.templates.renderer.constants.Constants.TD_CLASS`"
+    ":attr:`fastlife.templates.constants.Constants.TD_CLASS`"
   ] = None,
 #}
 <td
