+++ src/fastlife/components/Th.jinja
@@ -4,7 +4,7 @@
   class_: Annotated[
     str | None,
     "css class for the node, defaults to "
-    ":attr:`fastlife.templates.renderer.constants.Constants.TH_CLASS`."
+    ":attr:`fastlife.templates.constants.Constants.TH_CLASS`."
   ] = None,
 #}
 <th
