+++ src/fastlife/components/pydantic_form/Boolean.jinja
@@ -1,5 +1,5 @@
 {# doc Render a checkbox and a label for boolean field. #}
-{# def widget: Annotated[fastlife.templates.renderer.widgets.boolean.BooleanWidget, "widget to display."] #}
+{# def widget: Annotated[fastlife.adapters.jinjax.widgets.boolean.BooleanWidget, "widget to display."] #}
 <pydantic_form.Widget :widget="widget">
   <div class="pt-4">
     <div class="flex items-center">
