+++ src/fastlife/components/pydantic_form/Checklist.jinja
@@ -1,5 +1,5 @@
 {# doc Render a list of checkbox and a label for Literal and enum fields #}
-{# def widget: Annotated[fastlife.templates.renderer.widgets.checklist.ChecklistWidget, "widget to display."] #}
+{# def widget: Annotated[fastlife.adapters.jinjax.widgets.checklist.ChecklistWidget, "widget to display."] #}
 <pydantic_form.Widget :widget="widget">
   <div class="pt-4">
     <Details>
