+++ src/fastlife/components/pydantic_form/Hidden.jinja
@@ -1,3 +1,3 @@
 {# doc Render a hidden input field for a field. #}
-{# def widget: Annotated[fastlife.templates.renderer.widgets.hidden.HiddenWidget, "widget to display."] #}
+{# def widget: Annotated[fastlife.adapters.jinjax.widgets.hidden.HiddenWidget, "widget to display."] #}
 <Hidden :name="widget.name" :value="widget.value" :id="widget.id" />
