+++ src/fastlife/components/pydantic_form/Model.jinja
@@ -1,6 +1,6 @@
 {# doc Widget for pydantic BaseModel subclasses. #}
 {# def
-    widget: Annotated[fastlife.templates.renderer.widgets.model.ModelWidget, "widget to display."],
+    widget: Annotated[fastlife.adapters.jinjax.widgets.model.ModelWidget, "widget to display."],
     children_widget : Annotated[
         Sequence[fastlife.templating.renderer.widgets.base.Widget],
         "child widgets for every fields of the model."
