+++ src/fastlife/components/pydantic_form/Sequence.jinja
@@ -1,11 +1,11 @@
 {# doc Widget for pydantic BaseModel subclasses. #}
 {# def
-    widget: Annotated[fastlife.templates.renderer.widgets.sequence.SequenceWidget, "widget to display."],
+    widget: Annotated[fastlife.adapters.jinjax.widgets.sequence.SequenceWidget, "widget to display."],
     children_widgets : Annotated[
-        Sequence[fastlife.templates.renderer.widgets.base.Widget],
+        Sequence[fastlife.adapters.jinjax.widgets.base.Widget],
         "child widgets for every fields of the model."
     ],
-    type: Annotated[fastlife.templates.renderer.widgets.base.TypeWrapper, "child type wrapped."]
+    type: Annotated[fastlife.adapters.jinjax.widgets.base.TypeWrapper, "child type wrapped."]
 #}
 
 <pydantic_form.Widget :widget="widget">
