+++ src/fastlife/components/pydantic_form/Text.jinja
@@ -1,4 +1,4 @@
-{# def widget: Annotated[fastlife.templates.renderer.widgets.text.TextWidget, "widget to display."] #}
+{# def widget: Annotated[fastlife.adapters.jinjax.widgets.text.TextWidget, "widget to display."] #}
 
 <pydantic_form.Widget :widget="widget">
   <div class="pt-4">
