+++ src/fastlife/components/pydantic_form/Textarea.jinja
@@ -3,7 +3,7 @@ Render textarea widget for field of type text of event sequence.
 
 ::
 
-    from fastlife.templates.renderer.widgets.text import TextareaWidget
+    from fastlife.adapters.jinjax.widgets.text import TextareaWidget
     from pydantic import BaseModel, Field, field_validator
 
     class TagsForm(BaseModel):
@@ -19,7 +19,7 @@ Render textarea widget for field of type text of event sequence.
             return s.split() if s else []
 
  #}
-{# def widget: Annotated[fastlife.templates.renderer.widgets.text.TextareaWidget, "widget to display."] #}
+{# def widget: Annotated[fastlife.adapters.jinjax.widgets.text.TextareaWidget, "widget to display."] #}
 
 <pydantic_form.Widget :widget="widget">
   <div class="pt-4">
