+++ src/fastlife/components/pydantic_form/Widget.jinja
@@ -1,6 +1,6 @@
 {# doc Base component for widget #}
 {# def
-widget: Annotated[fastlife.templates.renderer.widgets.base.Widget, "widget to display."],
+widget: Annotated[fastlife.adapters.jinjax.widgets.base.Widget, "widget to display."],
 #}
 {% set container_id = widget.id + "-container" %}
 <div id="{{container_id}}">
