+++ src/fastlife/templates/__init__.py
@@ -3,10 +3,8 @@ Utilities for rendering HTML templates for page and components.
 """
 
 from .binding import Template, template
-from .renderer import JinjaxTemplateRenderer
 
 __all__ = [
-    "JinjaxTemplateRenderer",
     "Template",
     "template",
 ]
