+++ src/fastlife/templates/constants.py
