--- src/fastlife/templates/renderer/__init__.py
@@ -1,9 +1,0 @@
-"""Template renderer."""
-
-from .constants import Constants
-from .jinjax import JinjaxTemplateRenderer
-
-__all__ = [
-    "JinjaxTemplateRenderer",
-    "Constants",
-]
